*** empty log message ***
git-svn-id: http://svn.code.sf.net/p/gpac/code/trunk/gpac@530 63c20433-aa62-49bd-875c-5a186b69a8fb

+++ applications/osmo4_wx/menubtn.h
@@ -179,7 +179,7 @@ protected:
 
     virtual void CalcLayout(bool refresh);
 
-    int m_down;         // toggle state if m_down%2 then depressed
+    long m_down;         // toggle state if m_down%2 then depressed
     bool m_focused;     // mouse in window
     long m_button_style;
